+++ packages/keplr/src/wallet-connect/chain.ts
@@ -1,25 +1,26 @@
-import { ChainName, ChainWalletBase, State } from '@cosmos-kit/core';
+import { ChainRegistry, ChainWalletBase, State } from '@cosmos-kit/core';
+import { KeplrWalletConnectV1 } from '@keplr-wallet/wc-client';
 
 import { ChainWCKeplrData } from './types';
-import { WCKeplrWallet } from './wc-keplr';
+import { WCKeplrWallet } from './keplr';
+import { Keplr } from '@keplr-wallet/types';
 
 export class ChainWCKeplr extends ChainWalletBase<
+  KeplrWalletConnectV1,
   ChainWCKeplrData,
   WCKeplrWallet
 > {
 
-  constructor(_chainName: ChainName, keplrWallet: WCKeplrWallet) {
-    super(_chainName, keplrWallet);
+  constructor(_chainRegistry: ChainRegistry, keplrWallet: WCKeplrWallet) {
+    super(_chainRegistry, keplrWallet);
   }
 
   get client() {
-    return (async () => {
-      const client = await this.mainWallet.client;
-      if (!client) {
-        throw new Error('No WCKeplr installed!');
-      }
-      return client;
-    })();
+    return this.mainWallet.client;
+  }
+
+  get qrUri() {
+    return this.mainWallet.qrUri;
   }
 
   get username(): string | undefined {
@@ -38,7 +39,6 @@ export class ChainWCKeplr extends ChainWalletBase<
     } catch (e) {
       this.setState(State.Error);
       this.setMessage((e as Error).message);
-      throw e;
     }
   }
 }