+++ packages/keplr/src/wallet-connect/index.ts
@@ -1,0 +1,2 @@
+export * from './chain';
+export * from './keplr';