+++ packages/wallets/src/wc-keplr/index.ts
@@ -1,2 +1,0 @@
-export * from './chain';
-export * from './wc-keplr';